Underscore.jl

Underscore provides a single macro @_ to make it easier to pass closures to higher order functions. This is done by translating "placeholder expressions" containing placeholders _ or _1,_2,... . The key benefits of _ placeholders are

They avoid the need to come up with argument names (for example, the x in x->x+1 may not be meaningful).

Brevity. For example _+1.

API and Examples

Underscore.@_ — Macro

@_ func(ex1, [ex2 ...])

Convert ex1,ex2,... into anonymous functions when they have _ placeholders, and pass them along to func.

When multiple _ are present in a single sub-expression they become successive arguments to a single anonymous function. That is, @_ map(_+2^_, A) is equivalent to @_ map((x,y)->x+2^y, A).

Numbered placeholders _1,_2,... may be used if you need to reorder,repeat or omit arguments. For example @_ map(_2+_1, A, B) is equivalent to map((x,y)->(y,x), A, B).

Piping and composition chains are treated as a special case where the replacement recurses into sub-expressions. That is, the following two are equivalent:

@_ f1(ex1)  |>     f2(ex1)
@_(f1(ex1)) |>  @_(f2(ex1))

Examples

@_ can be very convenient for simple mapping operations in cases where broadcasting syntax is awkward. For example, to get the second last element of each array in a collection:

julia> @_ map(_[end-1],  [[1,2,3], [4,5]])
2-element Array{Int64,1}:
 2
 4

If you need to repeat an argument more than once the numbered form can be useful:

julia> @_ map(_1^_1,  [1,2,3])
3-element Array{Int64,1}:
  1
  4
 27

For manipulating tabular data @_ provides convenient syntax:

julia> data = [(x="a", y=1),
               (x="b", y=2),
               (x="c", y=3)];

julia> @_ filter(!startswith(_.x, "a"), data)
2-element Array{NamedTuple{(:x, :y),Tuple{String,Int64}},1}:
 (x = "b", y = 2)
 (x = "c", y = 3)

Combined with a lazy Map and Filter it gives super simple but convenient manipulation of tabular data:

julia> Filter(f) = x->filter(f,x);   Map(f) = x->map(f,x);

julia> @_ data |>
          Filter(!startswith(_.x, "a")) |>
          Map(_.y)
2-element Array{Int64,1}:
 2
 3

Design

We observe that anonymous functions are generally passed directly to another "outer" function. For example, in map(x->x*y, A) the outer function is map. Therefore one can place the @_ outside the map which leads to less visual clutter and improved clarity. This is particularly relevant because use of a macro in a function argument list tends to need additional parenentheses. Compare to the obvious alternatives:

    @_ map(_+1, A)
    map(@_(_+1), A)
    map(x->x+1, A)

With this "outermost-but-one" placement in mind, one can generalize to pipelines where anonymous functions are generally used as arguments to filter and map. This works in a particularly nice way for lazy versions of Filter and Map, allowing expressions such as

    @_  data         |>
        Filter(_>10) |>
        Map(_+1)

Somewhat of a design conundrum is how to make this work more natively with non-lazy filter and map. It might make sense to support double underscores to "escape an extra level", allowing things such as

    @_  data             |>
        filter(_>10, __) |>
        map(_+1, __)

to mean what is desired.