Underscores.jl

Underscores provides simple syntax for passing closures to functions with a macro @_. This enables use like @_ map(_+1, xs) and @_ people |> filter(_.age > 40, __) |> map(_.name, __).

Tutorial

Basic use of _

@_ and _ placeholders are for making functions to pass to other functions. This can be very convenient for simple uses of map in cases where broadcasting syntax is awkward. For example, to get the second last element of each array in a collection:

julia> @_ map(_[end-1],  [[1,2,3], [4,5]])
2-element Array{Int64,1}:
 2
 4

Repeated use of _ creates functions with multiple arguments. For example,

julia> @_ map("X $(repeat(_,_))", ["a","b","c"], [1,2,3])
3-element Array{String,1}:
 "X a"
 "X bb"
 "X ccc"

Numbered placeholders like _1 can be useful when you need to repeat arguments or reorder them. For example,

julia> @_ map("X $_2 $(repeat(_1,_2))", ["a","b","c"], [1,2,3])
3-element Array{String,1}:
 "X 1 a"
 "X 2 bb"
 "X 3 ccc"

Tabular data

@_ can be particularly helpful for manipulating tabular data, especially when combined with piping. Let's filter a list of named tuples:

julia> table = [(x="a", y=1),
                (x="b", y=2),
                (x="c", y=3)];

julia> @_ filter(!startswith(_.x, "a"), table)
2-element Array{NamedTuple{(:x, :y),Tuple{String,Int64}},1}:
 (x = "b", y = 2)
 (x = "c", y = 3)

@_ is especially useful when combined with double underscore placeholders __ and piping syntax. In the following, think of __ as the table, and _ as an individual row:

julia> @_ table |>
          filter(!startswith(_.x, "a"), __) |>
          map(_.y, __)
2-element Array{Int64,1}:
 2
 3

Reference

Underscores.@_ — Macro

@_ func(ex1, [ex2 ...])

Convert ex1,ex2,... into anonymous functions when they have _ placeholders, and pass them along to func.

When multiple _ are present in a single sub-expression they become successive arguments to a single anonymous function. That is, @_ map(_+2^_, A) is equivalent to @_ map((x,y)->x+2^y, A).

Numbered placeholders _1,_2,... may be used if you need to reorder,repeat or omit arguments. For example @_ map(_2+_1, A, B) is equivalent to map((x,y)->y+x, A, B).

Piping and composition chains are treated as a special case where the replacement recurses into sub-expressions. That is, the following two are equivalent:

@_ f1(ex1)  |>     f2(ex1)
@_(f1(ex1)) |>  @_(f2(ex1))

The placeholder __ (and numbered versions __1,__2,...) may be used to expand the closure scope to the whole expression. That is, the following are equivalent:

@_ func(a,__)
x->func(a,x)

Examples

@_ can be very convenient for simple mapping operations in cases where broadcasting syntax is awkward. For example, to get the second last element of each array in a collection:

julia> @_ map(_[end-1],  [[1,2,3], [4,5]])
2-element Array{Int64,1}:
 2
 4

If you need to repeat an argument more than once the numbered form can be useful:

julia> @_ map(_1^_1,  [1,2,3])
3-element Array{Int64,1}:
  1
  4
 27

For manipulating tabular data @_ provides convenient syntax:

julia> data = [(x="a", y=1),
               (x="b", y=2),
               (x="c", y=3)];

julia> @_ filter(!startswith(_.x, "a"), data)
2-element Array{NamedTuple{(:x, :y),Tuple{String,Int64}},1}:
 (x = "b", y = 2)
 (x = "c", y = 3)

It's especially useful when combined with double underscore placeholders __ and piping syntax. Think of __ as the table, and _ as an individual row:

julia> @_ data |>
          filter(!startswith(_.x, "a"), __) |>
          map(_.y, __)
2-element Array{Int64,1}:
 2
 3

Design

Underscore syntax for Julia has been discussed at great length in #24990, #5571 and elsewhere in the Julia community. The design for Underscores.jl grew out of this discussion. A great many packages have had a go at macros for this, including at least ChainMap.jl, ChainRecursive.jl, FunctionalData.jl, Hose.jl, Lazy.jl, LightQuery.jl, LambdaFn.jl, MagicUnderscores.jl, Pipe.jl, Query.jl and SplitApplyCombine.jl

The key benefits of _ placeholders are

They avoid the need to come up with argument names (for example, the x in x->x+1 may not be meaningful).

Brevity. For example _+1.

One design difficulty is that much of the package work has focussed on the piping and tabular data manipulation scenario. However as a language feature a compelling general solution for _ placeholders needs wider appeal.

Starting with the need to be useful outside of tabular data manipulation, we observe that anonymous functions are generally passed directly to another "outer" function. For example, in map(x->x*y, A) the outer function is map. However, putting @_ inside the function call leads to a lot of visual clutter, especially because it needs to be parenthesized to avoid consuming the remaining arguments to map. However, one can place the @_ on the function receiving the closure which results in less visual clutter and improved clarity. Compare:

@_ map(_+1, A)   # This design

map(@_(_+1), A)  # The obvious alternative

map(x->x+1, A)   # Current status quo

With this "outermost-but-one" placement in mind, one can generalize to pipelines where anonymous functions are generally used as arguments to filter and map. This works in a particularly nice way for lazy versions of Filter and Map, allowing expressions such as

Filter(f) = x->filter(f,x);   Map(f) = x->map(f,x);

@_  data         |>
    Filter(_>10) |>
    Map(_+1)

However, julia natively has non-lazy filter and map so we'd really like a way to make these directly useable as well. For this we introduce the longer placeholder __ to escape the extra function call to the outer level. This is also appealing because the larger data structure (ie, the full table) ends up being represented by __, while the smaller row data structure is the smaller placeholder _. Thus we get:

@_  data             |>
    filter(_>10, __) |>
    map(_+1, __)